backend/requirements.txt
daphne
Django==4.2.8
django-comment-migrate
django-cors-headers
django-filter
django-ranged-response
django-simple-captcha
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyMySQL
django-redis
drf-spectacular
drf-spectacular-sidecar
user-agents==2.2.0
openpyxl
requests
channels
channels_redis
#psycopg2